+++ include/bitcoin/bitcoin/chain/transaction_output.hpp
@@ -58,9 +58,9 @@ public:
 
     void reset();
 
-    size_t satoshi_size() const;
+    uint64_t satoshi_size() const;
 
-    static size_t satoshi_fixed_size();
+    static uint64_t satoshi_fixed_size();
 
 private:
